requirements.txt
streamlit
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
openai
unstructured
faiss-cpu
tiktoken
pydantic>=2.0.0,<3.0.0
